feat: answer vote 구현중

+++ server/src/main/java/pro/stackOverFlow/answer/dto/AnswerResponseDto.java
@@ -24,6 +24,6 @@ public class AnswerResponseDto {
 
     private String answerSelectionStatus;
 
-    private int voteCount;
+    private long voteCount;
 
 }
